V1.0 custom FAB component
Quick add front end mostly implemented. Two issues:
1) I can't figure out how to 'link' the datepicker and input field. Desired behaviour is: when a new date is chosen in the datepicker, the value in the input field updates accordingly.
2) This entire component needs to be linked to the back end (i.e. send values keyed in by user to the database)

One consideration: 
- Should the category input field be changed to a drop-down menu instead, consisting of currently existing categories + option to create a new category? 
OR keep the text input field? Then when the user types, it shows suggestions of existing categories (e.g. user types 'fo' and 'Food' is suggested). Eventually when the data is sent to the back end, if the category entered matches any of the currently existing categories in the database, it'll get added to that. Otherwise, it will create a new category (e.g. if user enters 'hhh' but only Food and Transport currently exist in the database, a new category hhh gets created)

Food jio screens haven't been implemented yet. I'll get to that next.